tried using a compact aos data structure instead of soa

+++ applicationLibrary/molecularDynamics/molecularDynamicsLibrary/MoleculeLJ.cpp
@@ -14,8 +14,8 @@ MoleculeLJ::MoleculeLJ(const std::array<double, 3> &pos, const std::array<double
 const std::array<double, 3> &MoleculeLJ::getOldF() const { return _oldF; }
 void MoleculeLJ::setOldF(const std::array<double, 3> &oldForce) { _oldF = oldForce; }
 
-size_t MoleculeLJ::getTypeId() const { return _typeId; }
-void MoleculeLJ::setTypeId(size_t typeId) { _typeId = typeId; }
+uint16_t MoleculeLJ::getTypeId() const { return _typeId; }
+void MoleculeLJ::setTypeId(uint16_t typeId) { _typeId = typeId; }
 
 size_t MoleculeLJ::getLiveId() const { return _liveId; }
 void MoleculeLJ::setLiveId(size_t liveId) { _liveId = liveId; }
